throw exception on failed request
If request to Odoo fails with faultCode, an exception should be thrown
to halt further processing.

Simply returning faultCode on error is not good, as there is no way to
distinguish the faultCode from expected returned id for example.

This might lead to unexpected consequences like assigning an id of 1 to
a record alltho it was faultCode 1.